open_ai/pyproject.toml
[tool.poetry]
name = "my_openai"
version = "0.1.0"
description = ""
authors = ["Wide Genesis <widegenesis@gmail.com>"]

readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
tenacity = "^8.4.0"
openai = "^1.40.1"
python-dotenv = "^1.0.1"
az_logger = {path = "../az_logger"}
settings = {path = "../settings"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
